+++ CHANGES.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file. This module adheres to [Semantic Versioning](https://semver.org/).
 ## 0.0.0
 - Add ValveController class
-- Add ValveControllerI2c class
+- Add ValveControllerI2c class for I2C support
 - Add RotaryController class
-- Add RotaryControllerTic class
+- Add RotaryControllerTic class for Tic stepper driver support
+- Add PumpController class
+- Add PumpControllerTic class for Tic stepper driver support